Port over clean-squashed to main cleanup script under the -s flag
The intent here is to run the clean-squashed flow over all local
branches that are _not_ cleaned up via the normal flow of merging.  As
this script cleans up the local branches and remotes that have been
merged in, the remaining branches are either:

       - WIP branches (should not be touched)
       - Stale branches that have been squash-merged in

For those remaining local branches, this will create dangling squashed
commits to see if we can remove them.  This will only affect local
branches and not cleanup any squash-merged remotes that have not
already been deleted (this is a TODO though).